removed classes from controllers
