--- CKODemoShop/Checkout/Bank.cs
@@ -1,8 +1,0 @@
-﻿namespace CKODemoShop.Checkout
-{
-    public class Bank : IIBank
-    {
-        public string Key { get; set; }
-        public string Value { get; set; }
-    }
-}
