+++ CKODemoShop/Checkout/IssuersResponse.cs
@@ -1,0 +1,10 @@
+﻿using Checkout.Common;
+using System.Collections.Generic;
+
+namespace CKODemoShop.Checkout
+{
+    public class IssuersResponse : Resource
+    {
+        public IList<IssuingCountry> Countries { get; set; }
+    }
+}
