+++ CKODemoShop/Checkout/IssuingBank.cs
@@ -1,0 +1,8 @@
+﻿namespace CKODemoShop.Checkout
+{
+    public class IssuingBank
+    {
+        public string Bic { get; set; }
+        public string Name { get; set; }
+    }
+}
