+++ CKODemoShop/Checkout/KlarnaProduct.cs
@@ -1,0 +1,20 @@
+﻿using Newtonsoft.Json;
+
+namespace CKODemoShop.Checkout
+{
+    public class KlarnaProduct
+    {
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
+        [JsonProperty(PropertyName = "quantity")]
+        public int Quantity { get; set; }
+        [JsonProperty(PropertyName = "unit_price")]
+        public int UnitPrice { get; set; }
+        [JsonProperty(PropertyName = "tax_rate")]
+        public int TaxRate { get; set; }
+        [JsonProperty(PropertyName = "total_amount")]
+        public int TotalAmount { get; set; }
+        [JsonProperty(PropertyName = "total_tax_amount")]
+        public int TotalTaxAmount { get; set; }
+    }
+}
