+++ CKODemoShop/Checkout/KlarnaSessionRequest.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace CKODemoShop.Checkout
+{
+    public class KlarnaSessionRequest
+    {
+        [JsonProperty(PropertyName = "purchase_country")]
+        public string PurchaseCountry { get; set; }
+        public string Currency { get; set; }
+        public string Locale { get; set; }
+        public int Amount { get; set; }
+        [JsonProperty(PropertyName = "tax_amount")]
+        public int TaxAmount { get; set; }
+        public IEnumerable<KlarnaProduct> Products { get; set; }
+    }
+}
