+++ CKODemoShop/Hubs/ITypedHubClient.cs
@@ -1,11 +1,11 @@
-﻿using CKODemoShop.Controllers;
+﻿using CKODemoShop.Checkout;
 using System.Threading.Tasks;
 
 namespace CKODemoShop.Hubs
 {
     public interface IWebhooksHubClient
     {
-        Task WebhookReceived(CheckoutWebhook checkoutWebhook);
+        Task WebhookReceived(Webhook checkoutWebhook);
         Task ConnectionMappedToPayment(string connectionId, string paymentId);
     }
 }
